move examples and tests to runtime directory
