--- instrumentation/runtime/example/doc.go
@@ -1,5 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-// Package main provides an example use of the runtime instrumentation.
-package main
