--- instrumentation/runtime/example/main.go
@@ -1,60 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-//go:build go1.18
-// +build go1.18
-
-package main
-
-import (
-	"context"
-	"log"
-	"os"
-	"os/signal"
-	"time"
-
-	"go.opentelemetry.io/contrib/instrumentation/runtime"
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/exporters/stdout/stdoutmetric"
-	"go.opentelemetry.io/otel/sdk/metric"
-	"go.opentelemetry.io/otel/sdk/resource"
-	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
-)
-
-var res = resource.NewWithAttributes(
-	semconv.SchemaURL,
-	semconv.ServiceName("runtime-instrumentation-example"),
-)
-
-func main() {
-	exp, err := stdoutmetric.New(stdoutmetric.WithPrettyPrint())
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// Register the exporter with an SDK via a periodic reader.
-	read := metric.NewPeriodicReader(
-		exp,
-		metric.WithInterval(1*time.Second),
-		metric.WithProducer(runtime.NewProducer()),
-	)
-	provider := metric.NewMeterProvider(metric.WithResource(res), metric.WithReader(read))
-	defer func() {
-		err := provider.Shutdown(context.Background())
-		if err != nil {
-			log.Fatal(err)
-		}
-	}()
-	otel.SetMeterProvider(provider)
-
-	ctx, cancel := signal.NotifyContext(context.Background(), os.Interrupt)
-	defer cancel()
-
-	log.Print("Starting runtime instrumentation:")
-	err = runtime.Start(runtime.WithMinimumReadMemStatsInterval(time.Second))
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	<-ctx.Done()
-}
